Allow configurable network names.
Some distros use a different name from "freenode". Allow the user to
configure it.  We'll provide a default here, and use a regex so that
as long as "freenode" appears in the name (and not in the name of any
other network the user is connected to, I hope), the default should
Just Work (TM).

Fixes #1